src/requirements.txt
# ── src/requirements.txt ─────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==23.0.0
azure-identity==1.14.1
azure-cosmos==4.6.0
cryptography==41.0.7
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl>=3.1.5
pdfplumber==0.10.4      # ← NEW (PDF parsing)
